can't wait to see these bugs in action

+++ src/main/java/web/Application.java
@@ -33,7 +33,7 @@ public class Application {
             Crop c = new Crop();
             c.setName("Tomato");
             c.setImageUrl("https://ui-ex.com/images/tomato-vector-svg-3.png");
-            c.setDiameter((float)2.0);
+            c.setDiameter(2);
             for(Crop crop : repository.findAll())
             {
                 if(crop.getId() == 2)
